refactor(templates): splits the `resources/{name}.ts` file to `data/{name}.ts` and `routes/api/{name}.ts` to avoid `Deno` from leaking into client
