+++ templates/crm/data/accounts.ts
@@ -1,24 +1,6 @@
 import { z } from "zod";
 import { faker } from "npm:@faker-js/faker@8.4.0";
-import { defineApiEndpoint } from "netzo/plugins/api/plugin.ts";
-import { DenoKvResource } from "netzo/plugins/api/resources/mod.ts";
-import { authenticate, log } from "netzo/plugins/api/hooks/mod.ts";
 import { ulid } from "netzo/plugins/api/utils.ts";
-
-export const accounts = defineApiEndpoint({
-  name: "accounts",
-  idField: "id",
-  resource: DenoKvResource({ prefix: ["accounts"] }),
-  hooks: {
-    all: [authenticate(), log()],
-    find: [],
-    get: [],
-    create: [],
-    update: [],
-    patch: [],
-    remove: [],
-  },
-});
 
 // schemas:
