+++ templates/crm/fresh.gen.ts
@@ -8,6 +8,12 @@ import * as $_app from "./routes/_app.tsx";
 import * as $_layout from "./routes/_layout.tsx";
 import * as $accounts_id_ from "./routes/accounts/[id].tsx";
 import * as $accounts_index from "./routes/accounts/index.tsx";
+import * as $api_accounts from "./routes/api/accounts.ts";
+import * as $api_contacts from "./routes/api/contacts.ts";
+import * as $api_deals from "./routes/api/deals.ts";
+import * as $api_interactions from "./routes/api/interactions.ts";
+import * as $api_invoices from "./routes/api/invoices.ts";
+import * as $api_transactions from "./routes/api/transactions.ts";
 import * as $contacts_id_ from "./routes/contacts/[id].tsx";
 import * as $contacts_index from "./routes/contacts/index.tsx";
 import * as $deals_index from "./routes/deals/index.tsx";
@@ -40,6 +46,12 @@ const manifest = {
     "./routes/_layout.tsx": $_layout,
     "./routes/accounts/[id].tsx": $accounts_id_,
     "./routes/accounts/index.tsx": $accounts_index,
+    "./routes/api/accounts.ts": $api_accounts,
+    "./routes/api/contacts.ts": $api_contacts,
+    "./routes/api/deals.ts": $api_deals,
+    "./routes/api/interactions.ts": $api_interactions,
+    "./routes/api/invoices.ts": $api_invoices,
+    "./routes/api/transactions.ts": $api_transactions,
     "./routes/contacts/[id].tsx": $contacts_id_,
     "./routes/contacts/index.tsx": $contacts_index,
     "./routes/deals/index.tsx": $deals_index,
