+++ templates/crm/islands/accounts/Form.tsx
@@ -1,8 +1,8 @@
 import { useSignal } from "@preact/signals";
 import { Button } from "netzo/components/button.tsx";
 import { createOnSubmit, Form } from "netzo/components/blocks/form/form.tsx";
-import { Account, accountSchema } from "@/resources/accounts.ts";
-// import { I18N } from "@/resources/accounts.ts";
+import { Account, accountSchema } from "@/data/accounts.ts";
+// import { I18N } from "@/data/accounts.ts";
 
 type FormProps = {
   data?: Account;
