+++ templates/crm/islands/accounts/Table.tsx
@@ -15,7 +15,7 @@ import {
 } from "netzo/components/avatar.tsx";
 import { IconCopy } from "netzo/components/icon-copy.tsx";
 import { Badge } from "netzo/components/badge.tsx";
-import { type Account, accountSchema, I18N } from "@/resources/accounts.ts";
+import { type Account, accountSchema, I18N } from "@/data/accounts.ts";
 
 // NOTE: define columns in island (route to island function serialization unsupported)
 export const getColumns = ({ options }: TableProps): TableProps["columns"] => [
