+++ templates/crm/islands/deals/Kanban.tsx
@@ -6,7 +6,7 @@ import {
   Kanban as _Kanban,
   type KanbanProps,
 } from "netzo/components/blocks/kanban/kanban.tsx";
-import { type Deal, I18N } from "@/resources/deals.ts";
+import { type Deal, I18N } from "@/data/deals.ts";
 
 // NOTE: define columns in island (route to island function serialization unsupported)
 export const getColumns = (_props: TableProps): TableProps["columns"] => [];
