+++ templates/crm/islands/interactions/Form.tsx
@@ -1,8 +1,8 @@
 import { useSignal } from "@preact/signals";
 import { Button } from "netzo/components/button.tsx";
 import { createOnSubmit, Form } from "netzo/components/blocks/form/form.tsx";
-import { Interaction, interactionSchema } from "@/resources/interactions.ts";
-// import { I18N } from "@/resources/interactions.ts";
+import { Interaction, interactionSchema } from "@/data/interactions.ts";
+// import { I18N } from "@/data/interactions.ts";
 
 type FormProps = {
   data?: Interaction;
