+++ templates/crm/islands/interactions/Table.tsx
@@ -14,7 +14,7 @@ import {
   I18N,
   type Interaction,
   interactionSchema,
-} from "@/resources/interactions.ts";
+} from "@/data/interactions.ts";
 
 // NOTE: define columns in island (route to island function serialization unsupported)
 export const getColumns = (
