+++ templates/crm/islands/invoices/Table.tsx
@@ -10,7 +10,7 @@ import { Grid } from "netzo/components/blocks/table/table.grid.tsx";
 import { toDate, toDateTime } from "netzo/components/blocks/format.ts";
 import { IconCopy } from "netzo/components/icon-copy.tsx";
 import { Badge } from "netzo/components/badge.tsx";
-import { I18N, type Invoice } from "@/resources/invoices.ts";
+import { I18N, type Invoice } from "@/data/invoices.ts";
 
 // NOTE: define columns in island (route to island function serialization unsupported)
 export const getColumns = ({ options }: TableProps): TableProps["columns"] => [
