+++ templates/crm/netzo.ts
@@ -1,10 +1,10 @@
 import { Netzo } from "netzo/mod.ts";
-import { accounts } from "./resources/accounts.ts";
-import { contacts } from "./resources/contacts.ts";
-import { deals } from "./resources/deals.ts";
-import { interactions } from "./resources/interactions.ts";
-import { invoices } from "./resources/invoices.ts";
-import { transactions } from "./resources/transactions.ts";
+import { accounts } from "./routes/api/accounts.ts";
+import { contacts } from "./routes/api/contacts.ts";
+import { deals } from "./routes/api/deals.ts";
+import { interactions } from "./routes/api/interactions.ts";
+import { invoices } from "./routes/api/invoices.ts";
+import { transactions } from "./routes/api/transactions.ts";
 
 export const netzo = await Netzo({
   auth: Deno.env.get("DENO_REGION") ? { providers: { netzo: {} } } : undefined,
