+++ templates/crm/routes/api/accounts.ts
@@ -1,0 +1,18 @@
+import { defineApiEndpoint } from "netzo/plugins/api/plugin.ts";
+import { DenoKvResource } from "netzo/plugins/api/resources/mod.ts";
+import { authenticate, log } from "netzo/plugins/api/hooks/mod.ts";
+
+export const accounts = defineApiEndpoint({
+  name: "accounts",
+  idField: "id",
+  resource: DenoKvResource({ prefix: ["accounts"] }),
+  hooks: {
+    all: [authenticate(), log()],
+    find: [],
+    get: [],
+    create: [],
+    update: [],
+    patch: [],
+    remove: [],
+  },
+});
