+++ templates/crm/routes/contacts/[id].tsx
@@ -1,6 +1,6 @@
 import { defineRoute } from "$fresh/server.ts";
 import { Separator } from "netzo/components/separator.tsx";
-import type { Contact } from "@/resources/contacts.ts";
+import type { Contact } from "@/data/contacts.ts";
 import { FormContact } from "@/islands/contacts/Form.tsx";
 import { netzo } from "@/netzo.ts";
