+++ templates/crm/routes/contacts/index.tsx
@@ -1,6 +1,6 @@
 import { defineRoute } from "$fresh/server.ts";
 import type { TableProps } from "netzo/components/blocks/table/use-table.ts";
-import { type Contact, I18N } from "@/resources/contacts.ts";
+import { type Contact, I18N } from "@/data/contacts.ts";
 import { Table } from "@/islands/contacts/Table.tsx";
 import { netzo } from "@/netzo.ts";
